Move logic for get categories out
