+++ frontend/presentations/pages/homepage/data-feed/homepage.json
@@ -1,4 +1,6 @@
 {
   "title": "Hai Bui Cool",
-  "heading": "Hai Bui oi co len"
+  "heading": "Hai Bui oi co len",
+  "categories": [],
+  "css": ""
 }
