+++ frontend/presentations/pages/homepage/homepageController.js
@@ -1,12 +1,9 @@
-const css = require("../../../src/css/css");
 const homepageService = require("./services/homepageService");
 
 let homepageController = async (req, res) => {
-	let data = await homepageService.buildData();
-	console.time("Test");
-
-	data.css = css.getFileContent("./assets/css/ifarmer-homepage-min.css");
-	console.timeEnd("Test");
+	console.time("Prepare data for homepage");
+	let data = await homepageService.prepareData();
+	console.timeEnd("Prepare data for homepage");
 	return res.render("pages/homepage/views/homepage", data);
 };
