+++ frontend/presentations/pages/homepage/services/homepageServiceSpec.js
@@ -1,10 +1,10 @@
 const homepageService = require("./homepageService");
 
 describe("homepageService", () => {
-	describe("buildData", () => {
+	describe("prepareData", () => {
 		it("should return correct data 1", () => {
 			let title = "Hai Bui";
-			let data = homepageService.buildData();
+			let data = homepageService.prepareData();
 			expect(title).toEqual(data.title);
 		});
 	});
